fix: bind controllers before custom connectedCallback behavior
If a controller synchronously dispatches events in its
connectedCallback, `data-action` handlers register the event too late,
because connectedCallback itself is a synchronous callback, while
`MutationObserver` is tasked.

This changes `bind` to be called _before_ custom `connectedCallback`
behavior so that any events dispatched in a `connectedCallback` are
captured by the `data-action` handler, making for a simpler
understanding of `data-action` handlers.

By making this change we allow developers to not have to internalise the
idea of dispatching events during Synchronous or Asynchronous
tasks/microtasks. Developers can simply dispatch an event without
worrying if it was synchronous or not.

Co-authored-by: Kristján Oddsson <koddsson@gmail.com>